+++ src/CfdiUtils/Elements/Retenciones10/Extranjero.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Extranjero extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'retenciones:Extranjero';
+    }
+}