+++ src/CfdiUtils/Elements/Retenciones10/Totales.php
@@ -1,0 +1,27 @@
+<?php
+namespace CfdiUtils\Elements\Retenciones10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Totales extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'retenciones:Totales';
+    }
+
+    public function addImpRetenidos(array $attributes = [], array $children = []): ImpRetenidos
+    {
+        $impRetenidos = new ImpRetenidos($attributes, $children);
+        $this->addChild($impRetenidos);
+        return $impRetenidos;
+    }
+
+    public function multiImpRetenidos(array ...$elementAttributes): self
+    {
+        foreach ($elementAttributes as $attributes) {
+            $this->addImpRetenidos($attributes);
+        }
+        return $this;
+    }
+}